student-management-system/pyproject.toml
[project]
name = "student-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator[email]>=2.3.0",
    "fastapi[standard]>=0.129.0",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.34",
]
